+++ packages/styles/src/packages/atomic/toast/rarui-toast.css.ts
@@ -7,9 +7,8 @@ const toast = recipe({
     position: "relative",
     display: "flex",
     alignItems: "center",
-    justifyContent: "space-between",
     boxSizing: "border-box",
-    gap: varsThemeBase.spacing["3xs"],
+    gap: varsThemeBase.spacing["2xs"],
     maxWidth: "100%",
     width: "420px",
     zIndex: varsThemeBase.zIndex[100],
@@ -55,13 +54,15 @@ const toast = recipe({
      */
     size: {
       small: {
-        padding: `${varsThemeBase.spacing["4xs"]} ${varsThemeBase.spacing["3xs"]}`,
+        padding: `${varsThemeBase.spacing["3xs"]} ${varsThemeBase.spacing["2xs"]}`,
+        gap: varsThemeBase.spacing["3xs"],
       },
       medium: {
-        padding: varsThemeBase.spacing["3xs"],
+        padding: `${varsThemeBase.spacing["2xs"]} ${varsThemeBase.spacing.xs}`,
       },
       large: {
-        padding: varsThemeBase.spacing.s,
+        padding: `${varsThemeBase.spacing.xs} ${varsThemeBase.spacing.s}`,
+        gap: varsThemeBase.spacing.xs,
       },
     },
     /**
@@ -207,8 +208,8 @@ const toastContainer = recipe({
     position: "fixed",
     display: "flex",
     flexDirection: "column",
-    gap: "16px",
-    padding: "16px",
+    gap: varsThemeBase.spacing.xs,
+    padding: varsThemeBase.spacing.xs,
     listStyle: "none",
   },
 
@@ -304,7 +305,7 @@ export const container = recipe({
     position: "fixed",
     flexDirection: "column-reverse",
     width: "fit-content",
-    gap: "1rem",
+    gap: varsThemeBase.spacing.xs,
   },
   variants: {
     placement: {
@@ -338,10 +339,15 @@ export const container = recipe({
   },
 });
 
+const iconContainer = style({
+  alignSelf: "self-start",
+});
+
 export const styles = {
   toast,
   container,
   content,
+  iconContainer,
   toastContainer,
   wrapper,
 };